move useIncrementSongPlayCount()
