+++ projects/graphql-client/src/index.ts
@@ -46,3 +46,4 @@ export * from "./remote/mutations/useRestorePassword"
 export * from "./remote/mutations/useSongFileDuplicates"
 export * from "./remote/mutations/useGenerateUploadableUrl"
 export * from "./remote/mutations/useSubmitSongFromRemoteFile"
+export * from "./remote/mutations/useIncrementSongPlayCount"
